# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Light For Change - Digital Candle Map API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats/countries:
    get:
      operationId: getCountryStats
      tags: [stats]
      summary: Candle count by country
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Country stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationStat"
  /stats/states:
    get:
      operationId: getStateStats
      tags: [stats]
      summary: Candle count by state
      parameters:
        - name: country
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: State stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationStat"
  /stats/cities:
    get:
      operationId: getCityStats
      tags: [stats]
      summary: Candle count by city
      parameters:
        - name: state
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: City stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationStat"
...
components:
  schemas:
...
    LocationStat:
      type: object
      properties:
        name:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required: [name, count, percentage]
...